ML pipeline/requirements.txt
# Core ML & Deep Learning
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Scientific Computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Model Utilities
joblib>=1.3.0

# API Clients
# API Clients
supabase>=2.27.0
redis>=5.0.0
redis>=5.0.0

# Event-Driven Architecture
fastapi>=0.110.0
uvicorn>=0.29.0
celery[redis]>=5.3.6

# NLP & Clustering
gliner>=0.1.0
hdbscan>=0.8.33
sumy>=0.11.0
supabase>=2.27.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
geopy>=2.4.0

# Utilities
python-dotenv>=1.0.0
